[refactor] (/user) 회원가입 및 로그인 기능 패키지 경로 수정
