+++ src/main/java/com/yoonNeun/MyDreamPartner/domain/user/application/UserLoginService.java
@@ -1,17 +1,16 @@
 package com.yoonNeun.MyDreamPartner.domain.user.application;
 
-import com.yoonNeun.MyDreamPartner.domain.oauth.application.OAuthService;
-import com.yoonNeun.MyDreamPartner.domain.oauth.domain.AuthToken;
-import com.yoonNeun.MyDreamPartner.domain.oauth.domain.AuthTokenGenerator;
-import com.yoonNeun.MyDreamPartner.domain.oauth.domain.OAuthResonse;
-import com.yoonNeun.MyDreamPartner.domain.user.domain.User;
+import com.yoonNeun.MyDreamPartner.domain.user.domain.dto.AuthToken;
+import com.yoonNeun.MyDreamPartner.domain.user.domain.AuthTokenGenerator;
+import com.yoonNeun.MyDreamPartner.domain.user.domain.dto.OAuthResonse;
+import com.yoonNeun.MyDreamPartner.domain.user.domain.entity.User;
 import com.yoonNeun.MyDreamPartner.domain.user.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class LoginService {
+public class UserLoginService {
 
     private final OAuthService oAuthService;
     private final UserRepository userRepository;
