+++ src/main/java/com/yoonNeun/MyDreamPartner/domain/user/application/UserSignUpService.java
@@ -1,7 +1,7 @@
 package com.yoonNeun.MyDreamPartner.domain.user.application;
 
 import com.yoonNeun.MyDreamPartner.domain.user.infrastructure.UserSignUpRepository;
-import com.yoonNeun.MyDreamPartner.domain.user.domain.UserProfile;
+import com.yoonNeun.MyDreamPartner.domain.user.domain.entity.UserProfile;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
